implemented updateMangerProfile
