+++ Modules/Manager/Http/Controllers/ManagerController.php
@@ -5,7 +5,6 @@ namespace Edaacil\Modules\Manager\Http\Controllers;
 
 use Edaacil\Modules\BaseController;
 use Edaacil\Modules\Manager\Http\Repositories\ManagerRepository;
-use Edaacil\Modules\Manager\Http\Requests\UpdateMangerProfile;
 
 class ManagerController extends BaseController
 {
@@ -31,11 +30,5 @@ class ManagerController extends BaseController
         return view($this->_config['view']);
     }
 
-    public function updateManagerInformation(UpdateMangerProfile $updateMangerProfile)
-    {
-        $manager = $this->managerRepository->updateManagerInformation($updateMangerProfile->all());
-        if ($manager)
-            session()->flash('success', 'Manager-Profile Updated Successfully');
-        return redirect()->back();
-    }
+
 }
